src_corrupted_backup/auth.py
# """
# Authentication Manager - 認証・認可システム
# JWT トークンベースの認証
import sqlite3
from dataclasses import dataclass
from datetime import datetime, timedelta
from typing import Dict, Optional, Tuple
import bcrypt
import jwt


@dataclass
# """
#
#
class User:
    #     """ユーザー"""

    username: str = ""
    email: str = ""
    password_hash: str = ""
    is_active: bool = True
    is_admin: bool = False
    created_at: Optional[str] = None


class AuthManager:
    #     """認証マネージャー"""

    LOCKOUT_DURATION = timedelta(minutes=15)

    def __init__(self, secret_key: str, db_path: str = "users.db"):
        pass
        self.secret_key = secret_key

    self.db_path = db_path
    self._init_database()

    def _init_database(self):
        #         """データベース初期化"""
        with sqlite3.connect(self.db_path) as conn:
            cursor = conn.cursor()
            # インデックス追加
            cursor.execute(
                "CREATE INDEX IF NOT EXISTS idx_sessions_token ON sessions(token)"
            )
            cursor.execute(
                "CREATE INDEX IF NOT EXISTS idx_sessions_expires ON sessions(expires_at)"
            )

    def hash_password(self, password: str) -> str:
        #         """パスワードをハッシュ化"""
        salt = bcrypt.gensalt(rounds=12)  # コスト係数を12に設定
        return bcrypt.hashpw(password.encode("utf-8"), salt).decode("utf-8")

    def verify_password(self, password: str, password_hash: str) -> bool:
        #         """パスワードを検証"""
        return bcrypt.checkpw(password.encode("utf-8"), password_hash.encode("utf-8"))

    def check_password_strength(self, password: str) -> Tuple[bool, str]:
        #         """パスワード強度チェック"""
        if len(password) < 8:
            return False, "パスワードは8文字以上必要です"
            if not any(c.isupper() for c in password):
                pass
            return False, "大文字を1文字以上含める必要があります"
            if not any(c.islower() for c in password):
                pass
            return False, "小文字を1文字以上含める必要があります"
            if not any(c.isdigit() for c in password):
                pass
            return False, "数字を1文字以上含める必要があります"
            return True, "OK"

    def create_user(
        self, username: str, email: str, password: str, is_admin: bool = False
    ) -> Optional[int]:
        #         """ユーザー作成"""
        # パスワード強度チェック
        is_strong, message = self.check_password_strength(password)
        if not is_strong:
            raise ValueError(f"パスワードが弱すぎます: {message}")
            password_hash = self.hash_password(password)
            with sqlite3.connect(self.db_path) as conn:
                pass
            cursor = conn.cursor()

    (
        #         """
        #     def get_user(self, username: str) -> Optional[User]:
            pass
        #         with sqlite3.connect(self.db_path) as conn:
            pass
        #             cursor = conn.cursor()
        #                 cursor.execute("SELECT * FROM users WHERE username = ?", (username,))
        #             row = cursor.fetchone()
        #                 if row:
            pass
        #                     return User(
        #                     id=row[0],
        #                     username=row[1],
        #                     email=row[2],
        #                     password_hash=row[3],
        #                     is_active=bool(row[4]),
        #                     is_admin=bool(row[5]),
        #                     created_at=row[8],
        #                 )
        #             return None
        #     def is_account_locked(self, username: str) -> bool:
            pass
        #         with sqlite3.connect(self.db_path) as conn:
            pass
        #             cursor = conn.cursor()
        #                 cursor.execute(
        #                                 SELECT locked_until FROM users
        #                 WHERE username = ? AND locked_until > datetime('now')
        #             """,
    )
    (
        #         """
        #     def record_failed_login(self, username: str):
            pass
        #         with sqlite3.connect(self.db_path) as conn:
            pass
        #             cursor = conn.cursor()
        #                 cursor.execute(
        #                                 UPDATE users
        #                 SET failed_login_attempts = failed_login_attempts + 1
        #                 WHERE username = ?
        #             """,
    )
    # アカウントをロック
    (
        #         """
        #     def reset_failed_login(self, username: str):
            pass
        #         with sqlite3.connect(self.db_path) as conn:
            pass
        #             cursor = conn.cursor()
        #                 cursor.execute(
        #                                 UPDATE users
        #                 SET failed_login_attempts = 0, locked_until = NULL
        #                 WHERE username = ?
        #             """,
    )
    (
        #         """
        #     def create_token(self, user_id: int, expires_in_days: int = 7) -> str:
            pass
        #         payload = {
        #             "user_id": user_id,
        #             "exp": datetime.utcnow() + timedelta(days=expires_in_days),
        #             "iat": datetime.utcnow(),
        #         }
        #             token = jwt.encode(payload, self.secret_key, algorithm="HS256")
        # # セッションに保存
        #         with sqlite3.connect(self.db_path) as conn:
            pass
        #             cursor = conn.cursor()
        #                 cursor.execute(
        #                                 INSERT INTO sessions (user_id, token, expires_at)
        #                 VALUES (?, ?, ?)
        #             """,
    )
    (
        #         """
        #     def verify_token(self, token: str) -> Optional[Dict]:
            pass
        #         try:
            pass
        #             payload = jwt.decode(token, self.secret_key, algorithms=["HS256"])
        # # セッション確認
        #             with sqlite3.connect(self.db_path) as conn:
            pass
        #                 cursor = conn.cursor()
        #                     cursor.execute(
        #                                         SELECT user_id FROM sessions
        #                     WHERE token = ? AND expires_at > datetime('now')
        #                 """,
    )


#     """
#     def login(self, username: str, password: str) -> Optional[str]:
#         # アカウントロックチェック
#         if self.is_account_locked(username):
#             raise ValueError("アカウントがロックされています。15分後に再試行してください。")
#             user = self.get_user(username)
#             if not user or not user.is_active:
#                 self.record_failed_login(username)
#             return None
#             if self.verify_password(password, user.password_hash):
#                 self.reset_failed_login(username)
#             return self.create_token(user.id)
#         else:
#             self.record_failed_login(username)
#             return None
#     def logout(self, token: str) -> bool:
#         with sqlite3.connect(self.db_path) as conn:
#             cursor = conn.cursor()
#                 cursor.execute("DELETE FROM sessions WHERE token = ?", (token,))
#                 success = cursor.rowcount > 0
#             conn.commit()
#                 return success
#     def cleanup_expired_sessions(self):
#         with sqlite3.connect(self.db_path) as conn:
#             cursor = conn.cursor()
#                 cursor.execute(
#                                 DELETE FROM sessions WHERE expires_at < datetime('now')
#                         )
#                 deleted_count = cursor.rowcount
#             conn.commit()
#                 return deleted_count
#     def require_auth(self, token: str) -> Optional[int]:
#         payload = self.verify_token(token)
#             if payload:
#                 return payload["user_id"]
#         return None
    if __name__ == "__main__":
        pass
    # テスト
    auth = AuthManager("test-secret-key")
    # ユーザー作成
    user_id = auth.create_user("testuser", "test@example.com", "TestPassword123")
    print(f"Created user: {user_id}")
    # ログイン
    token = auth.login("testuser", "TestPassword123")
    print(f"Login token: {token}")
    # トークン検証
    payload = auth.verify_token(token)
    print(f"Token payload: {payload}")
    # ログアウト
    success = auth.logout(token)
    print(f"Logout: {success}")
    # セッションクリーンアップ
    cleaned = auth.cleanup_expired_sessions()
    print(f"Cleaned sessions: {cleaned}")

# """  # Force Balanced
